WIP - introduce discount strategy introduce FiveForAmountOffer
