+++ exercise/kotlin/day13/src/main/kotlin/santamarket/model/ChristmasElf.kt
@@ -1,5 +1,6 @@
 package santamarket.model
 
+import santamarket.model.offer.FiveForAmountOffer
 import santamarket.model.offer.Offer
 import santamarket.model.offer.TenPercentOffer
 import santamarket.model.offer.TwoForAmountOffer
@@ -12,6 +13,7 @@ class ChristmasElf(private val catalog: SantamarketCatalog) {
             when (offerType) {
                 SpecialOfferType.TEN_PERCENT_DISCOUNT -> TenPercentOffer(offerType, product, argument)
                 SpecialOfferType.TWO_FOR_AMOUNT -> TwoForAmountOffer(offerType, product, argument)
+                SpecialOfferType.FIVE_FOR_AMOUNT -> FiveForAmountOffer(offerType, product, argument)
                 else -> Offer(offerType, product, argument)
             }
         offers[product] = offer
